normalizers/mongo-cdc-clickhouse/src/requirements.txt
confluent-kafka==2.5.3
PyYAML==6.0.2
python-dateutil==2.9.0.post0
